pyproject.toml
[project]
name = "edyrix-backend"
version = "0.1.0"
description = "Edyrix EdTech Platform — FastAPI Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "sqlalchemy[asyncio]==2.0.30",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "pydantic==2.7.1",
    "pydantic-settings==2.3.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "firebase-admin==6.5.0",
    "razorpay==1.4.1",
    "boto3==1.34.131",
    "redis==5.0.4",
    "resend==2.0.0",
    "PyMuPDF==1.24.4",
    "slowapi==0.1.9",
    "httpx==0.27.0",
]

[dependency-groups]
dev = [
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
